refactor: separate logic and ui in CoinTossGame #10
